requirements.txt
pyomo==6.8.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
PyYAML==6.0.2
